KAN-35: few fixes
